refactor: extract shared message parsers into dedicated module
Extract duplicate parsing logic from MessageItem.tsx and StreamingMessage.tsx
into a single source of truth at src/lib/message-parsers.ts.

Changes:
- Create src/lib/message-parsers.ts with all shared parsers:
  parseImageGenRequest, parseImageGenResult, parseBatchPlan,
  parseAllShowWidgets, parseShowWidget, computePartialWidgetKey,
  stripStructuredBlocks
- Add WeakMap-based parse cache to avoid re-parsing identical content
  during React re-renders (same string reference → cached result)
- Update MessageItem.tsx to import from shared module + re-export
  for backward compatibility with existing test imports
- Update StreamingMessage.tsx to import from shared module
- Remove ~200 lines of duplicated parser code

Performance benefits:
- Eliminates duplicate regex compilation across components
- WeakMap cache avoids re-parsing on re-renders with unchanged content
- Reduces bundle size by deduplicating identical parser logic
- Single source of truth reduces maintenance burden
