+++ src/components/chat/StreamingMessage.tsx
@@ -14,80 +14,16 @@ import { Shimmer } from '@/components/ai-elements/shimmer';
 import { ImageGenConfirmation } from './ImageGenConfirmation';
 import { BatchPlanInlinePreview } from './batch-image-gen/BatchPlanInlinePreview';
 import { WidgetRenderer } from './WidgetRenderer';
-import { parseAllShowWidgets, computePartialWidgetKey, MalformedWidgetNotice } from './MessageItem';
+import { MalformedWidgetNotice } from './MessageItem';
+import {
+  parseAllShowWidgets,
+  computePartialWidgetKey,
+  parseImageGenRequest,
+  parseBatchPlan,
+  stripStructuredBlocks,
+} from '@/lib/message-parsers';
 import { PENDING_KEY, buildReferenceImages } from '@/lib/image-ref-store';
-import type { PlannerOutput, MediaBlock } from '@/types';
-
-interface ImageGenRequest {
-  prompt: string;
-  aspectRatio: string;
-  resolution: string;
-  referenceImages?: string[];
-  useLastGenerated?: boolean;
-}
-
-function parseImageGenRequest(text: string): { beforeText: string; request: ImageGenRequest; afterText: string; rawBlock: string } | null {
-  const regex = /```image-gen-request\s*\n?([\s\S]*?)\n?\s*```/;
-  const match = text.match(regex);
-  if (!match) return null;
-  try {
-    let raw = match[1].trim();
-    let json: Record<string, unknown>;
-    try {
-      json = JSON.parse(raw);
-    } catch {
-      // Attempt to fix common model output issues: unescaped quotes in values
-      raw = raw.replace(/"prompt"\s*:\s*"([\s\S]*?)"\s*([,}])/g, (_m, val, tail) => {
-        const escaped = val.replace(/(?<!\\)"/g, '\\"');
-        return `"prompt": "${escaped}"${tail}`;
-      });
-      json = JSON.parse(raw);
-    }
-    const beforeText = text.slice(0, match.index).trim();
-    const afterText = text.slice((match.index || 0) + match[0].length).trim();
-    return {
-      beforeText,
-      request: {
-        prompt: String(json.prompt || ''),
-        aspectRatio: String(json.aspectRatio || '1:1'),
-        resolution: String(json.resolution || '1K'),
-        referenceImages: Array.isArray(json.referenceImages) ? json.referenceImages : undefined,
-        useLastGenerated: json.useLastGenerated === true,
-      },
-      afterText,
-      rawBlock: match[0], // full ```image-gen-request...``` block for exact matching
-    };
-  } catch {
-    return null;
-  }
-}
-
-function parseBatchPlan(text: string): { beforeText: string; plan: PlannerOutput; afterText: string } | null {
-  const regex = /```batch-plan\s*\n?([\s\S]*?)\n?\s*```/;
-  const match = text.match(regex);
-  if (!match) return null;
-  try {
-    const json = JSON.parse(match[1]);
-    const beforeText = text.slice(0, match.index).trim();
-    const afterText = text.slice((match.index || 0) + match[0].length).trim();
-    return {
-      beforeText,
-      plan: {
-        summary: json.summary || '',
-        items: Array.isArray(json.items) ? json.items.map((item: Record<string, unknown>) => ({
-          prompt: String(item.prompt || ''),
-          aspectRatio: String(item.aspectRatio || '1:1'),
-          resolution: String(item.resolution || '1K'),
-          tags: Array.isArray(item.tags) ? item.tags : [],
-          sourceRefs: Array.isArray(item.sourceRefs) ? item.sourceRefs : [],
-        })) : [],
-      },
-      afterText,
-    };
-  } catch {
-    return null;
-  }
-}
+import type { MediaBlock } from '@/types';
 
 interface ToolUseInfo {
   id: string;
@@ -549,11 +485,7 @@ export function StreamingMessage({
             if (hasImageGenBlock || hasBatchPlanBlock) return <Shimmer>{t('streaming.thinking')}</Shimmer>;
             return null;
           }
-          const stripped = content
-            .replace(/```image-gen-request[\s\S]*?```/g, '')
-            .replace(/```batch-plan[\s\S]*?```/g, '')
-            .replace(/```show-widget[\s\S]*?(```|$)/g, '')
-            .trim();
+          const stripped = stripStructuredBlocks(content);
           return stripped ? <MessageResponse>{stripped}</MessageResponse> : null;
         })()}
